# services/soup-identity/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Soup Identity API
  version: 1.0.0
  description: |
    Thin Soup identity microservice (v1 frozen).
    Trust rules: Soup never calls Jellyfin; no QC mailbox; multi-server is always `servers[]`.
servers:
  - url: http://localhost:8787
    description: Local docker-compose / `npm run dev`
...
security:
  - SoupBearer: []
paths:
...
  /v1/servers/{serverId}:
    put:
      tags: [plugin]
      summary: Register or update a Jellyfin server (plugin auth)
      security:
        - PluginBasic: []
      parameters:
        - $ref: "#/components/parameters/serverId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerUpsert"
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerRecord"
        "401":
          description: Invalid plugin credentials
    delete:
      tags: [plugin]
      summary: Delete server and its entitlements
      security:
        - PluginBasic: []
      parameters:
        - $ref: "#/components/parameters/serverId"
      responses:
        "204":
          description: Deleted
        "401":
          description: Invalid plugin credentials
        "404":
          description: Not found
  /v1/servers/{serverId}/entitlements/{email}:
    put:
      tags: [plugin]
      summary: Entitle a Google email to this server
      security:
        - PluginBasic: []
      parameters:
        - $ref: "#/components/parameters/serverId"
        - $ref: "#/components/parameters/email"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EntitlementUpsert"
      responses:
        "200":
          description: Upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntitlementRecord"
        "401":
          description: Invalid plugin credentials
        "404":
          description: Unknown server
    delete:
      tags: [plugin]
      summary: Revoke entitlement
      security:
        - PluginBasic: []
      parameters:
        - $ref: "#/components/parameters/serverId"
        - $ref: "#/components/parameters/email"
      responses:
        "204":
          description: Revoked
        "401":
          description: Invalid plugin credentials
        "404":
          description: Not found
  /v1/servers/{serverId}/entitlements/{email}/transport-grants:
    post:
      tags: [plugin]
      summary: Deposit one-time transport grant (Tailscale auth-key blob)
      description: >
        Encrypts material at rest, applies TTL, and invalidates prior unclaimed
        grants for the same (server_id, email). Plugin list omits material.
      security:
        - PluginBasic: []
      parameters:
        - $ref: "#/components/parameters/serverId"
        - $ref: "#/components/parameters/email"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransportGrantDeposit"
      responses:
        "201":
          description: Grant deposited (material omitted)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransportGrantRecord"
        "400":
          description: Invalid body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Invalid plugin credentials
        "404":
          description: Unknown server or entitlement
    get:
      tags: [plugin]
      summary: List transport grants (material omitted)
      security:
        - PluginBasic: []
      parameters:
        - $ref: "#/components/parameters/serverId"
        - $ref: "#/components/parameters/email"
      responses:
        "200":
          description: Grants without material
          content:
            application/json:
              schema:
                type: object
                required: [grants]
                properties:
                  grants:
                    type: array
                    items:
                      $ref: "#/components/schemas/TransportGrantRecord"
        "401":
          description: Invalid plugin credentials
        "404":
          description: Unknown server
  /v1/servers/{serverId}/entitlements/{email}/transport-grants/{grantId}:
    delete:
      tags: [plugin]
      summary: Revoke a transport grant
      security:
        - PluginBasic: []
      parameters:
        - $ref: "#/components/parameters/serverId"
        - $ref: "#/components/parameters/email"
        - name: grantId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Revoked
        "401":
          description: Invalid plugin credentials
        "404":
          description: Not found
...
components:
  securitySchemes:
    SoupBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived Soup access token from device-link or refresh
    PluginBasic:
      type: http
      scheme: basic
      description: plugin_id:plugin_secret issued at plugin setup / BOOTSTRAP_PLUGIN_*
  parameters:
...
    serverId:
      name: serverId
      in: path
      required: true
      schema:
        type: string
    email:
      name: email
      in: path
      required: true
      schema:
        type: string
  schemas:
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
...
    ServerUpsert:
      type: object
      required: [name, audience]
      properties:
        name:
          type: string
        base_url:
          type: string
        magic_dns:
          type: string
        audience:
          type: string
          description: JWT aud value the plugin expects
        metadata:
          type: object
          additionalProperties: true
    ServerRecord:
      type: object
      required: [server_id, name, audience, plugin_id]
      properties:
        server_id:
          type: string
        name:
          type: string
        base_url:
          type: string
          nullable: true
        magic_dns:
          type: string
          nullable: true
        audience:
          type: string
        plugin_id:
          type: string
        metadata:
          type: object
          additionalProperties: true
    EntitlementUpsert:
      type: object
      properties:
        display_name:
          type: string
        jellyfin_user_hint:
          type: string
        metadata:
          type: object
          additionalProperties: true
    EntitlementRecord:
      type: object
      required: [server_id, email]
      properties:
        server_id:
          type: string
        email:
          type: string
        display_name:
          type: string
          nullable: true
        jellyfin_user_hint:
          type: string
          nullable: true
        metadata:
          type: object
          additionalProperties: true
    TransportGrantDeposit:
      type: object
      required: [grant_type, material, ttl_seconds]
      properties:
        grant_type:
          type: string
          enum: [tailscale_auth_key]
        material:
          type: string
          description: Full tskey-auth-… (stored encrypted at rest)
        ttl_seconds:
          type: integer
          minimum: 60
          maximum: 86400
          description: >
            Soup grant TTL in seconds (60–86400). Should be ≤ Tailscale key
            expirySeconds so unused keys die with the mailbox entry.
        tailscale_key_id:
          type: string
          description: Tailscale key id for later DELETE without parsing material
        capabilities:
          type: object
          additionalProperties: true
          description: Optional mint flags echo for audit
        single_claim:
          type: boolean
          default: true
    TransportGrantRecord:
      type: object
      required: [id, grant_type, expires_at]
      properties:
        id:
          type: string
          format: uuid
        grant_type:
          type: string
        expires_at:
          type: string
          format: date-time
        claimed_at:
          type: string
          format: date-time
          nullable: true
        material:
          type: string
          description: Only returned once on claim; omitted from plugin list
        tailscale_key_id:
          type: string
          nullable: true
        capabilities:
          type: object
          additionalProperties: true